--- app/utils.py
@@ -1,164 +1,0 @@
-from __future__ import annotations
-
-import asyncio
-import io
-from contextlib import suppress
-from textwrap import shorten
-from typing import TYPE_CHECKING, Any, NamedTuple
-
-import discord
-
-from app.setup import config
-
-if TYPE_CHECKING:
-    from typing_extensions import TypeIs
-
-MAX_ATTACHMENT_SIZE = 67_108_864  # 64 MiB
-
-type Account = discord.User | discord.Member
-GuildTextChannel = discord.TextChannel | discord.Thread
-
-
-class MessageData(NamedTuple):
-    content: str
-    channel: discord.abc.MessageableChannel
-    attachments: list[discord.File]
-    skipped_attachments: int
-    reactions: dict[str | discord.Emoji, int]
-
-
-async def scrape_message_data(message: discord.Message) -> MessageData:
-    return MessageData(
-        message.content,
-        message.channel,
-        *await _get_attachments(message),
-        _get_reactions(message),
-    )
-
-
-async def _get_attachments(message: discord.Message) -> tuple[list[discord.File], int]:
-    if not message.attachments:
-        return [], 0
-
-    attachments: list[discord.File] = []
-    skipped_attachments = 0
-    for attachment in message.attachments:
-        if attachment.size > MAX_ATTACHMENT_SIZE:
-            skipped_attachments += 1
-            continue
-
-        fp = io.BytesIO(await attachment.read())
-        attachments.append(discord.File(fp, filename=attachment.filename))
-
-    return attachments, skipped_attachments
-
-
-def _get_reactions(message: discord.Message) -> dict[str | discord.Emoji, int]:
-    reactions: dict[str | discord.Emoji, int] = {}
-    for reaction in message.reactions:
-        if isinstance(emoji := reaction.emoji, discord.Emoji) and not emoji.is_usable():
-            continue
-        if isinstance(emoji, discord.PartialEmoji):
-            continue
-        reactions[emoji] = reaction.count
-    return reactions
-
-
-def _format_subtext(executor: discord.Member | None, msg_data: MessageData) -> str:
-    lines: list[str] = []
-    if reactions := msg_data.reactions.items():
-        lines.append("   ".join(f"{emoji} x{count}" for emoji, count in reactions))
-    if executor:
-        assert isinstance(msg_data.channel, GuildTextChannel)
-        lines.append(f"Moved from {msg_data.channel.mention} by {executor.mention}")
-    if skipped := msg_data.skipped_attachments:
-        lines.append(f"(skipped {skipped} large attachment(s))")
-    return "".join(f"\n-# {line}" for line in lines)
-
-
-async def get_or_create_webhook(
-    name: str, channel: discord.TextChannel | discord.ForumChannel
-) -> discord.Webhook:
-    webhooks = await channel.webhooks()
-    for webhook in webhooks:
-        if webhook.name == name:
-            if webhook.token is None:
-                await webhook.delete()
-            else:
-                return webhook
-
-    return await channel.create_webhook(name=name)
-
-
-async def move_message_via_webhook(
-    webhook: discord.Webhook,
-    message: discord.Message,
-    executor: discord.Member | None = None,
-    *,
-    thread: discord.abc.Snowflake = discord.utils.MISSING,
-    thread_name: str = discord.utils.MISSING,
-) -> discord.WebhookMessage:
-    msg_data = await scrape_message_data(message)
-
-    subtext = _format_subtext(executor, msg_data)
-    content, file = format_or_file(msg_data.content, template=f"{{}}{subtext}")
-    if file:
-        msg_data.attachments.append(file)
-        content += "\n-# (content attached)"
-
-    msg = await webhook.send(
-        content=content,
-        poll=message.poll or discord.utils.MISSING,
-        username=message.author.display_name,
-        avatar_url=message.author.display_avatar.url,
-        allowed_mentions=discord.AllowedMentions.none(),
-        files=msg_data.attachments,
-        thread=thread,
-        thread_name=thread_name,
-        wait=True,
-    )
-    await message.delete()
-    return msg
-
-
-def is_dm(account: Account) -> TypeIs[discord.User]:
-    return not isinstance(account, discord.Member)
-
-
-def is_mod(member: discord.Member) -> bool:
-    return member.get_role(config.MOD_ROLE_ID) is not None
-
-
-def is_helper(member: discord.Member) -> bool:
-    return member.get_role(config.HELPER_ROLE_ID) is not None
-
-
-def format_or_file(
-    message: str, *, template: str | None = None
-) -> tuple[str, discord.File | None]:
-    if template is None:
-        template = "{}"
-
-    if len(full_message := template.format(message)) > 2000:
-        return template.format(""), discord.File(
-            io.BytesIO(message.encode()), filename="content.md"
-        )
-    return full_message, None
-
-
-async def try_dm(account: Account, content: str, **extras: Any) -> None:
-    if account.bot:
-        return
-    try:
-        await account.send(content, **extras)
-    except discord.Forbidden:
-        print(f"Failed to DM {account} with: {shorten(content, width=50)}")
-
-
-async def remove_view_after_timeout(
-    message: discord.Message,
-    timeout: float = 30.0,  # noqa: ASYNC109
-) -> None:
-    await asyncio.sleep(timeout)
-    with suppress(discord.NotFound, discord.HTTPException):
-        await message.edit(view=None)
